fix avatars: move getUserAvatarUrl function to utils.ts for better organization and update imports across components
